requirements.txt
numpy
pandas
scikit-learn
nltk
indic-nlp-library
joblib
langdetect
flask
